Replace database repository
