+++ app/Models/Database.php
@@ -4,6 +4,7 @@ namespace Pterodactyl\Models;
 
 use Illuminate\Container\Container;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Support\Facades\DB;
 use Pterodactyl\Contracts\Extensions\HashidsInterface;
 
 /**
@@ -28,10 +29,7 @@ class Database extends Model
      */
     public const RESOURCE_NAME = 'server_database';
 
-    /**
-     * The table associated with the model.
-     */
-    protected $table = 'databases';
+    public const DEFAULT_CONNECTION_NAME = 'dynamic';
 
     /**
      * The attributes excluded from the model's JSON form.
@@ -106,5 +104,74 @@ class Database extends Model
     public function server(): BelongsTo
     {
         return $this->belongsTo(Server::class);
+    }
+
+    /**
+     * Run the provided statement against the database on a given connection.
+     */
+    private function run(string $statement): bool
+    {
+        return DB::connection($this->connection)->statement($statement);
+    }
+
+    /**
+     * Create a new database on a given connection.
+     */
+    public function createDatabase(string $database): bool
+    {
+        return $this->run(sprintf('CREATE DATABASE IF NOT EXISTS `%s`', $database));
+    }
+
+    /**
+     * Create a new database user on a given connection.
+     */
+    public function createUser(string $username, string $remote, string $password, ?int $max_connections): bool
+    {
+        $args = [$username, $remote, $password];
+        $command = 'CREATE USER `%s`@`%s` IDENTIFIED BY \'%s\'';
+
+        if (!empty($max_connections)) {
+            $args[] = $max_connections;
+            $command .= ' WITH MAX_USER_CONNECTIONS %s';
+        }
+
+        return $this->run(sprintf($command, ...$args));
+    }
+
+    /**
+     * Give a specific user access to a given database.
+     */
+    public function assignUserToDatabase(string $database, string $username, string $remote): bool
+    {
+        return $this->run(sprintf(
+            'GRANT SELECT, INSERT, UPDATE, DELETE, CREATE, DROP, ALTER, REFERENCES, INDEX, LOCK TABLES, CREATE ROUTINE, ALTER ROUTINE, EXECUTE, CREATE TEMPORARY TABLES, CREATE VIEW, SHOW VIEW, EVENT, TRIGGER ON `%s`.* TO `%s`@`%s`',
+            $database,
+            $username,
+            $remote
+        ));
+    }
+
+    /**
+     * Flush the privileges for a given connection.
+     */
+    public function flush(): bool
+    {
+        return $this->run('FLUSH PRIVILEGES');
+    }
+
+    /**
+     * Drop a given database on a specific connection.
+     */
+    public function dropDatabase(string $database): bool
+    {
+        return $this->run(sprintf('DROP DATABASE IF EXISTS `%s`', $database));
+    }
+
+    /**
+     * Drop a given user on a specific connection.
+     */
+    public function dropUser(string $username, string $remote): bool
+    {
+        return $this->run(sprintf('DROP USER IF EXISTS `%s`@`%s`', $username, $remote));
     }
 }
